% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{batch.correct}
\alias{batch.correct}
\title{Perform Batch Correction of EEMs}
\usage{
batch.correct(eem.dir)
}
\arguments{
\item{\code{eem.dir}}{- The filepath to the top directory housing raw EEM files, as .dat files
produced by a Flouromax F4.}
}
\value{
Corrected EEMs as CSVs saved in grouped subdirectories of a "corrected" subdirectory created
 in the specified input directory. An index file is also saved for the EEMs in the "corrected"
 folder. See \code{\link{make.indicies}} for details.
}
\details{
Given all input files and prarameters, this code performs EEM corrections as in the
orignial Matlab code. Only a corrected EEM is output, other functions should be invoked to plot the
 correctred EEM or to calculate indicies. Function assumes un-altered F4 files have not been
 altered to remove headers.
}
\author{
Robert Lee
}
