% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_plot.R
\name{batch.plot}
\alias{batch.plot}
\title{Perform Bulk EEM Plotting}
\usage{
batch.plot(eem.dir, rayliegh.mask = T, white.mask = T)
}
\arguments{
\item{\code{eem.dir}}{- A directory containing previosuly corrected EEMs, generated by the \code{f4.eem.correct}
function. Will search recursively through the directory for corrected EEMs. Corrected EEMs have file names ending in "_c.csv"}

\item{\code{rayliegh.mask}}{- Should 1st and 2nd order Rayliegh scattering be masked? Defaults to TRUE.}

\item{\code{white.mask}}{- Should masked values be drawn white? Defaults to TRUE.}
}
\value{
EEM plots are saved in a plots folder, inside the directory given to \code{eem.dir}.
}
\details{
When a folder containing corrected EEMs \emph{somewhere} in it is passed to the function,
plots for all found EEMs are saved. The folder structure inside the directory is replicated in a
'plots' directory inside the given EEM folder\cr
\cr
}
\examples{

}
\author{
Robert Lee
}
