% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_parafac.R
\docType{data}
\name{cut.th}
\alias{cut.th}
\title{Perform PARAFAC Modeling on Corrected EEMs}
\format{An object of class \code{numeric} of length 1.}
\usage{
cut.th
}
\arguments{
\item{\code{}}{-}
}
\value{

}
\details{

}
\author{

}
\keyword{datasets}
