% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solid_grid_plot.R
\name{eem.grid.plot}
\alias{eem.grid.plot}
\title{Generate a Grid of Solid State Corrected EEM Plots}
\usage{
eem.grid.plot(corr.dir, combo.file, white.mask = T)
}
\arguments{
\item{\code{corr.dir}}{- The directory where corrected EEMs live}

\item{\code{combo.file}}{- The spreadsheet (Excel or CSV) of desired combinations. Reuired columns:
\itemize{
\item{"Name of Raw EEM"}
\item{"Corrected Name"}
\item{"Notes Column"}
}}
}
\value{
Files are saved to the \code{corr.dir} location specefied.
}
\details{
- This function takes specified correction combinations (from previously corrected solid
state eems, see \code{\link{solid.combo.correct}}) on a given base EEM, and plots all the
combinations specified in a grid.
}
\author{
Robert Lee

Maggie Bowman
}
