% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eem.R
\name{eem.plot}
\alias{eem.plot}
\title{Plot a corrected EEM}
\usage{
eem.plot(corr.eem, sample.name, save.dir)
}
\arguments{
\item{\code{corr.eem}}{- A corrected EEM}

\item{\code{sample.name}}{- The name associated with the sample}

\item{\code{save.dir}}{- The directory to save the EEM plot to.}
}
\value{

}
\details{
- Given a corrected EEM, the desired sample name, and where the output
plot should be saved, a plotted EEM is returned.
}
\author{
Robert Lee
}
