% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_eem.R
\name{expand.eem}
\alias{expand.eem}
\title{Expand an EEM into a 1 nm resolution size}
\usage{
expand.eem(eem)
}
\arguments{
\item{\code{eem}}{- An eem with row names and column names
corresponding to the excitation/emission wavelengths. If the eem has been corrected by the fluoro
package, then it can be loaded with the \code{read.corr.eem} function.}
}
\value{
Returns the expanded EEM as a data frame
}
\details{
For an EEM of any whole nanometer resolution, perform a spline interpolation on both
dimensions of the EEM to expand the EEM into a 1 nm resolution data frame. Warning- this function
differs slightly from the interp2 function used in original Matlab code.
}
\author{
Robert Lee
}
