% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3_eem_correct.R
\name{f3.eem.correct}
\alias{f3.eem.correct}
\title{Perform PARAFAC Modeling on Corrected EEMs}
\usage{
f3.eem.correct(eem.file, blank.file, abs.file, raman.file, save.name, save.dir,
  dil.fact = 1, uv.path.length = 1, em.cor.file, ex.cor.file, ram.cor.file)
}
\arguments{
\item{\code{eem.file}}{- The filepath to the raw EEM file, a .dat file produced by a Flouromax F4.}

\item{\code{blank.file}}{- The filepath to the blank file corresponding to the raw EEM.
A .dat file produced by a Flouromax F4.}

\item{\code{uv.file}}{- The filpath to the UV-Vis abosrbance spectra for the raw eem being
corrected.}

\item{\code{raman.file}}{- The filepath to the blank file corresponding to the raw EEM.
A .dat file produced by a Flouromax F4.}

\item{\code{save.name}}{- The desired filename of the corrected EEM.}

\item{\code{save.dir}}{- Where the corrected eem (as a CSV) will be saved.}

\item{\code{dil.fact}}{- Optional, defaults to 1 (no dillution). Only anter a value reflecting your
dilution factor if the sample has been dilluted.}

\item{\code{uv.path.length}}{- Optional, defaults to 1 cm. Only enter this value if a different
cuvette size was used.}
}
\value{
A the corrected EEM, and a CSV of the EEM is also saved to the specified save directory.
}
\details{
Given all input files and prarameters, this code performs EEM corrections as in the
orignial Matlab code. Only a corrected EEM is output, other functions should be invoked to plot the
 correctred EEM or to calculate indicies. Function assumes un-altered F4 files have not been
 altered to remove headers.
}
\examples{
\dontrun{
abs.file=system.file("extdata", "abs.csv", package = "fluoro")
blank.file=system.file("extdata", "blank.dat", package = "fluoro")
eem.file=system.file("extdata", "eem.dat", package = "fluoro")
raman.file=system.file("extdata", "raman.dat", package = "fluoro")

corr.eem=f4.eem.correct(eem.file, blank.file, abs.file, raman.file,
save.dir=tempdir(), save.name="corrected_eem")
}

#export - do when done

}
\author{
Robert Lee
}
