% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3_inst_corr.R
\name{f3.inst.corr}
\alias{f3.inst.corr}
\title{Perform F3 Instrument Correction on F3 Files}
\usage{
f3.inst.corr(f3.eem, f3.blank, em.corr.file, ex.corr.file, f3.raman, corr.fact,
  ramman.corr.file)
}
\arguments{
\item{\code{f3.eem}}{- An F3 EEM read in by the \code{\link{read.f3.eem}} function}

\item{\code{f3.blank}}{- An F3 Blank file read in by the \code{\link{read.f3.eem}} function}

\item{\code{em.corr.file}}{- The path to the emission correction file specific to the F3 used. Must be an XLS file.}

\item{\code{em.corr.file}}{- The path to the emission correction file specific to the F3 used. Must be an XLS file.}

\item{\code{f3.raman}}{-  An
F3 raman file read in by the \code{\link{read.f3.eem}} function}

\item{\code{corr.fact}}{- The
correction factor specific to the F3 used.}

\item{\code{ramman.corr.file}}{- The path to the Raman correction file specific to the F3 used. Must be an XLS file.}
}
\value{
A list of 3 objects:
\itemize{
\item{\code{ic.eem}}: The instrument corrected EEM
\item{\code{ic.blank}}: The instrument corrected Blank
\item{\code{ic.raman}}: The instrument corrected Raman.
}
}
\details{
Given a raw EEM, Blank and (optionally) Raman file from an F3 instrument,
instrument corrected versions are returned for use in correcting raw F3 EEMs.
Raman files are optional only to accomodate the correction of solid state EEMs-
\strong{raman correction information must be supplied when correcting an EEM of
aqueous samples!}
}
\author{
Robert Lee
}
