% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresh_calc.R
\name{fresh.calc}
\alias{fresh.calc}
\title{Perform PARAFAC Modeling on Corrected EEMs}
\usage{
fresh.calc(corr.eem)
}
\arguments{
\item{\code{corr.eem}}{- A previosuly corrected EEM, generated by the \code{f4.eem.correct}
function. If the EEM isn't loaded, read it in with the \code{read.corr.eem} function.}
}
\value{
The freshenss index value
}
\details{
Will expand an EEM to 1nm incriements in both directions,
and returns the calculated freshness index.
}
\author{
Robrt Lee
}
