% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hix_calc.R
\name{hix.calc}
\alias{hix.calc}
\title{Calculate the Humification Index (HIX)}
\usage{
hix.calc(corr.eem)
}
\arguments{
\item{\code{corr.eem}}{- A previosuly corrected EEM, generated by the \code{f4.eem.correct}
function. If the EEM isn't loaded, read it in with the \code{read.corr.eem} function.}
}
\value{
The calculated HIX value.
}
\details{
Given a corrected EEM produced by the \code{f4.eem.correct} function, the function will
expand the EEM to a 1 nm resolution matrix using a spline interpolation called on first the emission
and then the exictation dimension. The HIX calculation is then performed according to the Ohno
method (2002). \cr
\cr
NOTE: the original Matlab code does not rely on a two step interploation, but performs a bivariate
spline intrpolation using the \code{interp2} function. As a result, the HIX values produced by
this code an the MatLab \code{F4CorrectFunII} code begin to differ after the fourth decimal.
Reconciling the two code bases to produce identical HIX values is an ongoing effort.
}
\examples{

}
\author{
Robert Lee, adapting original Matlab code by...
}
