% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parafac.model.R
\name{parafac.model}
\alias{parafac.model}
\title{Perform PARAFAC Modeling on Corrected EEMs}
\usage{
parafac.model(expanded.eem, num.factors)
}
\arguments{
\item{\code{}}{-}
}
\value{

}
\details{

}
\author{

}
