% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raman_mask.R
\name{rayleigh.mask}
\alias{rayleigh.mask}
\title{Mask Rayleigh Scatter}
\usage{
rayleigh.mask(eem, mask.value = 0)
}
\arguments{
\item{\code{eem}}{- A properly-formatted EEM, produced or read by a \code{fluoro} function.}

\item{\code{mask.value}}{- Optional, defaults to zero. The numeric value that values in the mask area should be forced to.
\strong{NA/NaN/NULL values will prevent ploting in \code{fluoro}!}}
}
\value{
A masked EEM with dimensions and names preserved.
}
\details{
Masks first and second order Rayleigh Scatter on the supplied EEM with the specified value
(defaulting to 0 if no value entered.)
}
\author{
Robert Lee
}
