% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corr_eem.R
\name{read.corr.eem}
\alias{read.corr.eem}
\title{Read in a Corrected EEM}
\usage{
read.corr.eem(file)
}
\arguments{
\item{\code{file}}{- A character path pointing to the corrected eem file.
File should be a CSV, typically output by the f4.eem.correct function.}
}
\value{
A properly formatted EEM, as a data frame.
}
\details{
This file reads in the corrected EEMs produced by "f4.eem.correct",
and correctly formats it for use with other functions in the 'fluoro' package.
}
\author{
Robert Lee
}
