% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_f3_eem.R
\name{read.f3.eem}
\alias{read.f3.eem}
\title{Read in an EEM File Generated by a Fluoromax F3}
\usage{
read.f3.eem(f3.eem.file)
}
\arguments{
\item{\code{f3.eem.file}}{- A character path pointing to the EEM file.
File must be an xls file output by the Fluoromax software.}
}
\value{
A properly formatted EEM, as a data frame.
}
\details{
This function reads in raw EEMs produced by the F3, as an XLS file.
A correctly formatted EEM is returned for use with other functions in the 'fluoro' package.
}
\author{
Robert Lee
}
