% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_eem.R
\name{read.raw.eem}
\alias{read.raw.eem}
\title{Load a Raw EEM}
\usage{
read.raw.eem(file)
}
\arguments{
\item{\code{file}}{- The file path to the raw eem file. Should be tab-separated data.}
}
\value{
An EEM in the format required by \code{fluoro}.
}
\details{
Loads a raw (uncorrected) EEM into R for use with other \code{fluoro} functions.
}
\author{
Robert Lee
}
