% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solid_combo_correct.R
\name{solid.combo.correct}
\alias{solid.combo.correct}
\title{Batch Correcrt Solid State EEMs with Specifierd Combinations}
\usage{
solid.combo.correct(eem.dir, combo.file)
}
\arguments{
\item{\code{corr.dir}}{- The directory where corrected EEMs live}

\item{\code{combo.file}}{- The spreadsheet (Excel or CSV) of desired combinations. Required columns:
\itemize{
\item{"Name of Raw EEM"}
\item{"Corrected Name"}
\item{"Notes Column"}
}}
}
\value{
Corrected EEMs are saved with concatenated file names a "corrected" subdirectory of the
path passed to \code{eem.dir}.\cr{}
The direct output to the parent environment is a list with objects \code{input} and \code{corr.dir}.
These outputs exist only for passing to another batch function of solid state EEMs, and can be
ignored if not needed.
}
\details{
- This function takes specified correction combinationsc specified in the file passed to
 \code{combo.file}, and performs the outlined corrections.
}
\author{
Robert Lee

Maggie Bowman
}
