% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solid_indicies.R
\name{solid.indicies}
\alias{solid.indicies}
\title{Make a File of Summary Indicies Used in Analyzing Solid-state EEMs}
\usage{
solid.indicies(corr.eem, sample, save.dir)
}
\arguments{
\item{\code{corr.eem}}{- A corrected eem, either the output of \code{f4.eem.correct} or \code{read.corr.eem}.}

\item{\code{sample}}{- The sample name associated with the EEM.}

\item{\code{save.dir}}{- Where the index file ("indicies.csv") will be saved.}
}
\value{
A CSV of the calculated indicies, with column names and the sample ID corresponding to the indicies.
 If a file of indicies already exists in the save.dir, the values for the current EEM will be appended.
}
\details{
Given all input files and prarameters, this code produces summary indicies for the corrected
eem. The specifically, maximium emission at 370 nm excitation,
the fluorescence Index (FI), Humification Index, and Freshness Index are returned in the output file.
For specifics of on the calculation of FI, HIX, Freshness, see the following help files:
\enumerate{
  \item \code{\link{fi.calc}}
  \item \code{\link{hix.calc}}
  \item \code{\link{fresh.calc}}
}
}
\examples{
\dontrun{
#Save an example index file to the current working directory
corr.eem=fluoro::read.corr.eem(file=system.file("extdata", "corr_eem.csv", package = "fluoro"))

fluoro::solid.indicies(uv.file=uv.file, sample="sample_name", save.dir=getwd())
}

}
\author{
Robert Lee
}
