% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_eem.R
\name{subtract.eem}
\alias{subtract.eem}
\title{Subtract two Equal-Dimension EEMs}
\usage{
subtract.eem(p.eem, s.eem, neg.to.zero = FALSE)
}
\arguments{
\item{\code{p.eem}}{- An EEM to be subtracted from, either a raw or corrected EEM.}

\item{\code{s.eem}}{- An EEM to subtract from \code{p.eem}, either a raw or corrected EEM.}

\item{\code{neg.to.zero}}{- Optional, defaults to FALSE. IF set to TRUE, negative values are
coerced to 0.}
}
\value{
Corrected EEMs as CSVs saved in grouped subdirectories of a "corrected" subdirectory created
 in the specified input directory. An index file is also saved for the EEMs in the "corrected"
 folder. See \code{\link{make.indicies}} for details.
}
\details{
Perfoms quailty checks before performing straight subtraction of the 's.eem' input from
'p.eem' (p.eem=s.eem). If the 'neg.to.zero' parameter is set to TRUE, then negative values are
coerced to 0
}
\author{
Robert Lee
}
